Add broken vector flameball functionality

+++ src/com/tsb/main/GameObject.java
@@ -7,7 +7,7 @@ public abstract class GameObject {
 
     protected int x, y;
     protected ID id;
-    protected int velX, velY;
+    protected float velX, velY;
 
     public GameObject(int x, int y, ID id){
         this.x = x;
@@ -46,19 +46,19 @@ public abstract class GameObject {
         return id;
     }
 
-    public int getVelX() {
+    public float getVelX() {
         return velX;
     }
 
-    public void setVelX(int velX) {
+    public void setVelX(float velX) {
         this.velX = velX;
     }
 
-    public int getVelY() {
+    public float getVelY() {
         return velY;
     }
 
-    public void setVelY(int velY){
+    public void setVelY(float velY){
             this.velY = velY;
     }
